Structured-Report-Generation/pyproject.toml
[project]
name = "structured-report-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.54",
    "langchain-groq>=0.3.2",
    "langchain-nvidia-ai-endpoints>=0.3.9",
    "langgraph>=0.3.31",
    "langsmith>=0.3.32",
    "mcp-use>=1.2.7",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.5.4",
    "typing-extensions>=4.13.2",
]
